Move path builder class to utils.py
