+++ src/onepasswordconnectsdk/async_client.py
@@ -4,9 +4,8 @@ from httpx import HTTPError
 import json
 import os
 
-from onepasswordconnectsdk.connect import PathBuilder
 from onepasswordconnectsdk.serializer import Serializer
-from onepasswordconnectsdk.utils import build_headers, is_valid_uuid
+from onepasswordconnectsdk.utils import build_headers, is_valid_uuid, PathBuilder
 from onepasswordconnectsdk.errors import (
     FailedToRetrieveItemException,
     FailedToRetrieveVaultException,
